+++ src/App.js
@@ -1,8 +1,8 @@
 import { QueryClient, QueryClientProvider } from 'react-query'
 import { Routes, Route } from 'react-router-dom'
 import Hero from './sections/Hero'
-import SmartVaults from './pages/SmartVaults'
-import Primitives from './pages/Primitives'
+import Home from './pages/Home'
+import SmartVault from './pages/SmartVault'
 
 
 const queryClient = new QueryClient()
@@ -11,9 +11,8 @@ function App() {
   return (
     <QueryClientProvider client={queryClient}>
       <Routes>
-        <Route path="/" element={<SmartVaults />} />
-        <Route path="/smart-vaults" element={<SmartVaults />} />
-        <Route path="/smart-vaults/:id" element={<Primitives />} />
+        <Route path="/" element={<Home />} />
+        <Route path="/smart-vaults/:id" element={<SmartVault />} />
       </Routes>
     </QueryClientProvider>
   )
