+++ src/pages/SmartVault.js
